+++ templates/tem_admin_profile_4.php
@@ -44,18 +44,20 @@
 <div id="navbar">
             <a href="tem_admin_profile.php">Teachers Info</a>
             <a href="tem_admin_profile_2.php">Edit Teachers</a>
-            <a href="tem_admin_profile_3.php">Add</a>
-            <a href="tem_admin_profile_4.php">Remove Teacher</a>
+            <a href="tem_admin_profile_3.php">Add Teacher</a>
+            <a style="background-color: #2bff00;" href="tem_admin_profile_4.php">Remove Teacher</a>
             <a href="tem_admin_profile_5.php">Add Courses to Teacher</a>
+            <a href="tem_admin_profile_6.php">Remove Courses from Teacher</a>
+
         
         </div>
     
   <div class="new_content">
     
   <form action="tem_admin_profile_4.php" method="POST">
-      <input type="text" name="teacher_id">
+      <input type="text" name="teacher_id" placeholder="Teacher Id">
       <button style="margin-top: 50px;" type="submit" class="sub_btn"
-                                name="REGISTER">REGISTER</button>
+                                name="REGISTER">Remove Teacher</button>
       <!-- <button type="submit" class="sub_btn" name="search">SEARCH</button> -->
 
   </form>
@@ -75,19 +77,26 @@
         
         $value = $_POST['teacher_id'];
 
-        $sql = "delete from teacher_info where teacher_id='$value'";
+        $sql = "select * from teacher_info where teacher_id='$value';";
+        $result = mysqli_query($conn,$sql);
+        $count = mysqli_num_rows($result);
 
 
 
 
-        if(mysqli_query($conn, $sql))
+        if($count > 0)
         {
-          
-            echo "SUCCESS";
+           $sql = "delete from teacher_info where teacher_id='$value';";
+           mysqli_query($conn,$sql);
+
+            echo "<P style='color: green'> Teacher with Id : $value Removed Successfully </p>";
+            // header('Location: templates/tem_admin_profile.php');
           
         }
         else{
-          echo "ERROR : ".mysqli_error($conn);
+          echo "<P style='color: red'><b> No Teacher with Id : $value </b></p>";
+
+          // echo "ERROR : ".mysqli_error($conn);
         }
